Move most filesystem related storage code into new filestore_fs.py.
